Added a Scene class together with a SceneView.
The idea is that a developer simply registers different views to the current
activity in order to display different things on the screen. The SceneView
is (obviously) the 3D scene that should be displayed and connects directly to
an instance of engine.model.Scene. Other views will be added as the engine is
developed.