Implemented usage of vertex attributes during rendering.
It is now possible to register the usage of a specific vertex attribute to a
program. As a consequence, it's possible to use certain vertex attributes only
in certain programs.

This is done by looping through all available vertex attributes and checking
if they have been chosen to be registered in a certain program. If they have,
the attributes in question will be sent to the shaders in that program.