Added initial support for uniforms.
While vertex attributes are only part of a PrimitiveData instance, uniforms can
be either global or local to a primitive. In practice, this means that uniforms
can be defined either in a PrimitiveData instance or in a Program instance.
This way, it is possible to choose if a uniform should be global or local.