Added support for frame update listeners and surface change listeners.
A frame listener is just what it sounds like; a callback that will be called
whenever a frame is updated. The callback receives the time elapsed since last
frame.

A surface change listener is called whenever the underlying surface/display
changes. It is also called if a surface change listener is registered on an
already stable surface.