Converted dinja_bomb.x3d to a proper (working) format.
The previous object was using a mix of quads and triangles, something that is
not supported in OpenGL ES 2.0. The bomb object has now been converted and
only uses ordinary triangles (GL_TRIANGLES).

As I was updating the object anyway, I also took the opportunity to redo the
texture used on the bomb.