Added support for texturing.
To accomodate this change, support for samplers (a type of uniform) was added.
Right now, only textures in the format RGB565 are supported. Support for more
formats will be added as they are "needed". One limitation right now is
of course the lack of alpha channels.

The reading of X3D files has been improved and vertices are now duplicated
whenever the same vertex points to different texture coordinates. This makes
sure that we get correct texture mapping under OpenGL even when the UV
coordinates are per-face (like in X3D) and not per-vertex.

For most objects there is a significant sharing of UV coordinates between
vertices. With the dinja bomb test object the number of vertices are nearly
doubled from 86 to 168. If we were to duplicate all vertices we would have
gotten 504 vertices (the number of faces * 3). So we are definitely still
saving vertices.