Implemented support for frame time listeners in the OpenGL ES layer.
An object can implement IFrameTimeListener and register itself as a frame time
listener. It will then receive a call with the time it took to draw the last
frame. This can be used to for example measure the theoretical number of frames
per second that a device can achieve. This is often much more useful than the
real frames per second, because it's not capped to the refresh rate.