Added DebugView class and usage of it in DragWorldActivity.
This view takes a list of models that it should print out debug information
about. It then registers a DebugController that updates the view depending on
the values in the model.

Right now, it's possible to print out debug information about a scene, model
or camera.

If you try to pass an object that can't be debugged by the debug view, an
InvalidArgumentException is thrown.