Implemented support for transparency and blending.
When a texture is set using Mesh.setTexture(), the blending function is
selected automatically. The OpenGL ES layer then fetches this information and
takes action accordingly.

It is possible to control and override the blending that setTexture() selects
by calling PrimitiveData.setBlendFactor().

Also, it's possible to completely ignore the alpha channel of a texture by
calling Texture.setHasAlpha(false) before setting the texture on the mesh.