Created an initial implementation of a recursive "scene graph".
This implementation differs from some other scene graph implementations. For
example, the transformation (model) matrix is baked directly into groups. Thus,
there is no special transformation node.

Everything is a node. Even the scene is considered a node.

Right now, the group class or grouping is not yet implemented. Neither is deep
matrix multiplication. The implementation functions exactly as the previous
one did. The difference is that we are now working with general nodes instead
of meshes.