Transformations are now propagated up in the scene graph.
Maybe not the most efficient propagation right now, but it works exactly as
intended.

Even the model matrix in the scene can be used and is propagated up in the
scene graph. It can be used to translate, scale and rotate the whole scene.

Note that the camera is an exception and is not affected by the model matrix
in the scene.