Implemented the ability to rotate around a specified center other than 0,0,0.
This functionality can be reached by calling Node.getRotator().setCenter() and
is especially usefull together with groups of objects where a rotation center
often needs to be set.

The rotation point is specified in model space and is local to the node.
This means the rotation center is not affected by the translation of the node.