Began reworking the OpenGL ES layer.
There are a few problems with the current design. It has a number of cyclic
package-level dependencies and some of it doesn't feel very clean.

Instead of having a big wrapper towards the native bindings (the OpenGLES2
class), we will implement that functionality in separate classes that will call
OpenGLES2Native.