Implemented pixel precise picking of Meshes.
In the OpenGL ES layer this was accomplished by implementing a SelectionDraw
class that draws primitves in different colours using a custom vertex shader
and pixel shader. Classes that want to fetch what primitive was clicked simply
register themselves as listeners of this information.

In the actual engine, the FingerFlingMeshInputListener interface was
implemented. This interface allows controllers to receive a list of meshes
that the user "flinged" over. This can be used safely together with any of the
other input listeners.