Created a Frustum class for the camera and optimized matrix propagation.
The frustum class holds information about things such as near/far plane and
field of view. It also holds the projection matrix for the camera that it is
part of.

The view and projection matrices were being multiplied every single frame;
something that is completely unnecessary. Also, the viewProjectionMatrix was
being added in the end of the propagation instead of being added in the
begining, resulting in additional matrix operations.