Added the IFramePeekListener interface and used it in the SelectionDraw class.
This interface is used by drawers when they want to "peek" (read) the drawable
area they are drawing on. This was implemented to exclude peeking from any
frame rate calculation. Otherwise, functions such as glReadPixels() interfere
with that calculation as they stall and wait for any previous gl draw operation
to complete.