Added a Material class.
This class is part of Meshes and holds information about the material that a
mesh has. A material definition with the following "lighting model" was chosen:

1. Ambient intensity (how much ambient color the material attracts)
2. Diffuse color
3. Specular color
4. Shininess

To this, meshes can define vertex colours. These that are mixed in with the
material properties during rendering by the shader.

This is not yet used in any meaningful way, but will be used by the lighting
model once completed.