Added getNodeClosestToPoint() methods to the BaseNode class.
These will be documented better later, but this is how they work:

Two versions are available:
INode getNodeClosestToPoint(Point3f point)
INode getNodeClosestToPoint(Point3f point, List<INode> nodes)

Both versions find (as the name suggests) the node closest to the supplied
point. The closest point is found recursively and the algorithm always stops
at a leaf node. Note that, at each step of recursion, the algorithm only
enters (and searches recursively) the node that is the closest one. This means
that the closest node that the algorithm finds is not necessarily the closest
global leaf node.

The second version of the method is a static version that can take an arbitrary
list of nodes to search through, which can be useful whenever you want to be
specific.